+++ src/Rule/CharacterClass.php
@@ -90,7 +90,7 @@ abstract class CharacterClass implements RuleInterface
         $constraint = $this->getViolation($count);
 
         if ($constraint !== null) {
-            throw new RuleException($this, $this->getMessage($constraint, $count));
+            throw new RuleException($this, $constraint->getWeight(), $this->getMessage($constraint, $count));
         }
     }