+++ src/Rule/Length.php
@@ -71,7 +71,7 @@ final class Length implements RuleInterface
         $constraint = $this->getViolation($count);
 
         if ($constraint !== null) {
-            throw new RuleException($this, $this->getMessage($constraint, $count));
+            throw new RuleException($this, $constraint->getWeight(), $this->getMessage($constraint, $count));
         }
     }