+++ src/Rule/RuleException.php
@@ -18,13 +18,20 @@ final class RuleException extends RuntimeException
     private $rule;
 
     /**
+     * @var int Weight of violated constraint.
+     */
+    private $weight;
+
+    /**
      * @param RuleInterface $rule Rule that could not be enforced.
+     * @param int $weight Weight of violated constraint.
      * @param string $message Exception message.
      * @param Throwable|null $previous Previous exception, used for exception chaining.
      */
-    public function __construct(RuleInterface $rule, string $message, ?Throwable $previous = null)
+    public function __construct(RuleInterface $rule, int $weight, string $message, ?Throwable $previous = null)
     {
         $this->rule = $rule;
+        $this->weight = $weight;
 
         parent::__construct($message, /*code*/0, $previous);
     }
@@ -35,5 +42,13 @@ final class RuleException extends RuntimeException
     public function getRule(): RuleInterface
     {
         return $this->rule;
+    }
+
+    /**
+     * @return int Weight of violated constraint.
+     */
+    public function getWeight(): int
+    {
+        return $this->weight;
     }
 }